feat: refactor icon into data package, add parameter support and comments to icons
