+++ src/main/java/dev/dfonline/flint/actiondump/codeblocks/ActionType.java
@@ -1,6 +1,6 @@
 package dev.dfonline.flint.actiondump.codeblocks;
 
-import dev.dfonline.flint.actiondump.Icon;
+import dev.dfonline.flint.actiondump.data.Icon;
 
 public record ActionType(
         String name,
