+++ src/main/java/dev/dfonline/flint/actiondump/codeblocks/BlockTagType.java
@@ -1,6 +1,6 @@
 package dev.dfonline.flint.actiondump.codeblocks;
 
-import dev.dfonline.flint.actiondump.Icon;
+import dev.dfonline.flint.actiondump.data.Icon;
 
 public record BlockTagType(
         String name,
