+++ src/main/java/dev/dfonline/flint/actiondump/data/ParameterElementType.java
@@ -1,0 +1,10 @@
+package dev.dfonline.flint.actiondump.data;
+
+public record ParameterElementType(
+        String type,
+        boolean plural,
+        boolean optional,
+        String[] description,
+        String[] notes
+) {
+}
