+++ src/main/java/dev/dfonline/flint/actiondump/gamevalues/GameValueType.java
@@ -1,6 +1,6 @@
 package dev.dfonline.flint.actiondump.gamevalues;
 
-import dev.dfonline.flint.actiondump.Icon;
+import dev.dfonline.flint.actiondump.data.Icon;
 
 public record GameValueType(
         Icon icon,
